requirements/local.txt
# requirements/local.txt
-r base.txt
django-debug-toolbar>=4.0.0
pytest>=7.3.1
pytest-django>=4.5.2
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0